Transcode: Differentiate between local and remote streaming
This adds provisions for the concept of local and remote streaming. For local streaming bitrate is generally not a concern so higher quality transcoding can be used.
VAAPI will only be used for local streaming as the quality is not good enough at low bitrates for remote streaming compared to software encoding.

Signed-off-by: Scott Ware <scott2.ware@gmail.com>