Media Scanning: Implemented more robust and comprehensive metadata parsing.
Changed the file parser to FFProbe which is able to output JSON which is much easier to parse.
Media streams are now stored in separate tables which enables more detailed information to be stored for each type of stream.
The type of files in a media folder is now automatically determined during scanning.
Playlists are now detected as part of a normal media scan. Playlist content parsing remains an isolated task.
Changed Media Elements and Media Folders to use UUIDs.

Signed-off-by: Scott Ware <scott2.ware@gmail.com>